added leaderboard fixed finish

+++ Assets/Scripts/Core/GameTime.cs
@@ -24,7 +24,7 @@ namespace VitaliyNULL.Core
             return string.Format($"{_minutes}:{_seconds}");
         }
 
-        public int GetTime() => _allTime;
+        public float GetTime() => _allTime;
 
         public void SetTime(int minutes, int seconds)
         {
